Create python_ex
The COVID-19 pandemic has significantly impacted society in many different aspects, one of the most notable being the workforce. In March 2020, most companies were forced to implement remote work procedures and, having grown used to this new way of life, many elected not to return to the office even once the pandemic was declared to be over. San Francisco is one example of a prominent American city facing the ramifications of this post-pandemic digital era as office vacancy rates continue to increase, while rent prices continue to fall. This paper will examine data and trends to answer how the city of San Francisco has been impacted by the COVID-19 pandemic in terms of the labor market and demand for office space. Specifically, I will test the hypothesis that the increase in office vacancy rates is affecting business activity and public transit ridership rates. As less people commute into the city, more businesses will close.